NER tagging
Added NER tagging. Can create and then use the neural network. Uses the
FinBERT model and is trained on the turku-one corpus. All of this is in
the tagger.py - module. Also added some checks when creating a new
neural network that the destination folders are empty and that a new
network cannot be trained if the destination folders are not empty.