Handles longer texts when doing sentiment analysis
Splits the tokens in a way that each token sequence is a max of 510 + [CLS] and [SEP] tags. If the sequence is less than 512 long, it will be padded with 0s. After this the chunks are stacked and fed to the model. The average of the output is calculated and given to the input.