Added ner tag identification to post title and body analysis.
Checks the text and finds the tags from it. Then the tags are combined to dict in which the tags are combined to single words using B- and I- tags. Problem with words that are split by tokenizer so one word is identified as B- multiple times. This leads to some wierd outputs in the table.