Analysis to each sentiment tag and the sentiment of the context
Added the functionality to see how each found sentiment tag is corresponding to the the sentiment from which it was found from. Ex. finding Suomi from sentence Suomi on hieno maa, will result in finding Suomi as the ner tag and it is connected with the sentiment of the sentence.